+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 202x-xx-xx
 
 ### Added
-- 
+- Allow to use a custom class for the chart rendering via the `Chart(..., renderer_cls=MyClass)` keyword argument (see `examples/custom-renderer-class.py` for inspiration)
 
 ### Changed
 - Fixed off-by-one shift when rendering empty lines on the Y-axis (the issue was visible only when the Y-axis was on the left side)